Added command upper/lower test.
