+++ src/test/java/org/smack/application/CliApplicationTypeTest.java
@@ -5,66 +5,13 @@ import static org.junit.Assert.assertTrue;
 
 import java.io.ByteArrayOutputStream;
 import java.io.File;
-import java.io.IOException;
 import java.io.PrintStream;
-import java.io.StringReader;
-import java.util.List;
 
 import org.junit.Test;
-import org.smack.util.FileUtil;
 import org.smack.util.StringUtil;
 
 public class CliApplicationTypeTest
 {
-    @Test
-    public void TestHelp() throws IOException
-    {
-        PrintStream originalErrOut =
-                System.err;
-        ByteArrayOutputStream errOs =
-                new ByteArrayOutputStream();
-        System.setErr( new PrintStream( errOs ) );
-
-        PrintStream originalOut =
-                System.out;
-        ByteArrayOutputStream outOs =
-                new ByteArrayOutputStream();
-        System.setOut( new PrintStream( outOs ) );
-
-        ApplicationUnderTest.main( new String[0] );
-
-        System.err.flush();
-        System.setErr( originalErrOut );
-        System.out.flush();
-        System.setOut( originalOut );
-
-        String expectedString =
-                "ApplicationUnderTest\n" +
-                        "\n" +
-                        "The following commands are supported:\n" +
-                        "\n" +
-                        "cmdBoolean: boolean\n" +
-                        "cmdByte: byte\n" +
-                        "cmdDouble: double\n" +
-                        "cmdEnum: [FRIDAY, MONDAY, SATURDAY, SUNDAY, THURSDAY, TUESDAY, WEDNESDAY]\n" +
-                        "cmdFile: File\n" +
-                        "cmdFloat: float\n" +
-                        "cmdInt: int\n" +
-                        "cmdLong: long\n" +
-                        "cmdShort: short\n" +
-                        "cmdString: String\n\n";
-        List<String> expectedLines =
-                FileUtil.readLines(
-                        new StringReader( expectedString ) );
-        List<String> receivedLines =
-                FileUtil.readLines(
-                        new StringReader( errOs.toString() ) );
-
-        assertEquals(
-                expectedLines,
-                receivedLines );
-    }
-
     private String hex( Number n )
     {
         return String.format(
